Multipart forms improvements (WIP)
